moving myFavorites up in state to ApplicationViews
